Use faraday 2.x explicitly
opensearch-transport 2.0.1 and 1.0.1 unlocks using faraday 2.x but
fluent-plugin-opensearch isn't ready for it, as a result out_opensearch
doesn't work on newly install.
This patch resolves it by using faraday 2.x explicitly.
Another way is lock to use faraday 1.x but staying on older one might
not be a good idea.

Fix #70

Signed-off-by: Takuro Ashie <ashie@clear-code.com>